Create autoSell.js
to add this for school chromebooks, copy the code by hitting the two folders, than you need to favorite the page, than hit more. in the URL section, type 'javascript:'
than hit control v or paste the script that you put. Hit done and click on that when you are on the blooket market and booom complete.